+++ models/metrics/metrics.go
@@ -1,5 +1,7 @@
 package metrics
 
+import "errors"
+
 const (
 	ReportType = "reports"
 	ReportStr  = "Report"
@@ -156,6 +158,66 @@ type ThresholdCrossingTopNRequest struct {
 	Granularity        string `json:"granularity,omitempty"`
 	Timeout            int32  `json:"timeout,omitempty"`
 	NumResults         int32  `json:"numResults,omitempty"`
+}
+
+type TopNForMetric struct {
+	// One of the two must be populated for the request to be valid, domains or monitoredObjects.
+	// But if both are given, then the behaviour will be the query will be based on a subset of monitoredObjects that belong to the domains.
+	// List of domains (optional)
+	Domains []string `json:"domains,omitempty"`
+	// List of monitored objects (optional)
+	MonitoredObjects []string `json:"monitoredObjects,omitempty"`
+
+	// Required Time range for the requestin ISO 8601 format for intervals
+	Interval string `json:"interval,,omitempty"`
+	// Rquired Vendor (to avoid overlaps, eg: flowmeter does not have Jitter values
+	// so if you do a min TopN then you'll just get a list of 0s)
+	TenantID string `json:"tenant,omitempty"`
+	// Timeout for the request
+	Timeout int32 `json:"timeout,omitempty"`
+	// Number of Results (default is 10)
+	NumResult int32 `json:"NumResults,omitempty"`
+
+	// Operation - 'avg', 'min', 'max'
+	Aggregator string `json:"aggregator,omitempty"`
+	// Metric that we are apply Aggregation to
+	Metric []MetricIdentifier `json:"metrics,omitempty"`
+
+	// Metrics that are related and interesting BUT are NOT part of the post aggregation
+	MetricsView []MetricAggregation `json:"metricsView,omitempty"`
+}
+
+type MetricAggregation struct {
+	// Metric name eg jitterP95
+	Metric string `json:"metric,omitempty"`
+	// Operation - 'sum', 'count', 'min', 'max'
+	Aggregator string `json:"aggregator,omitempty"`
+	// Name for this Aggregation (must be unique)
+	Name string `json:"name,omitempty"`
+}
+
+func (tpn *TopNForMetric) Validate() (*TopNForMetric, error) {
+	req := tpn
+	if req.Timeout == 0 {
+		req.Timeout = 5000
+	}
+
+	if len(tpn.Metric) == 0 {
+		return nil, errors.New("Metric must not be empty")
+	}
+	if tpn.NumResult == 0 {
+		tpn.NumResult = 10
+	}
+
+	if len(req.TenantID) == 0 {
+		return nil, errors.New("Tenant must not be empty.")
+	}
+
+	if len(req.Interval) == 0 {
+		return nil, errors.New("Interval must not be empty")
+	}
+
+	return req, nil
 }
 
 type ThresholdCrossingRequest struct {